FP:258: Changed inventory transaction field type.
git-svn-id: svn://svn.code.sf.net/p/floreantpos/code/trunk@1708 37e33ebe-f127-4a38-aa86-8919ee84a29d

+++ src/com/floreantpos/model/InventoryItem.java
@@ -65,7 +65,7 @@ public class InventoryItem extends BaseInventoryItem {
 			
 			inventoryItem.setName(strings[index++]);
 			inventoryItem.setUnitPerPackage(POSUtil.parseDouble(strings[index++]));
-			inventoryItem.setTotalPackages(POSUtil.parseInteger(strings[index++]));
+			inventoryItem.setTotalPackages(POSUtil.parseDouble(strings[index++]));
 			inventoryItem.setAveragePackagePrice(POSUtil.parseDouble(strings[index++]));
 			inventoryItem.setTotalRecepieUnits(POSUtil.parseDouble(strings[index++]));
 			inventoryItem.setUnitPurchasePrice(POSUtil.parseDouble(strings[index++]));
